requirements.txt
pymupdf4llm == 0.0.17
python-docx == 1.1.2
